Provide separate open and process session calls
The current start() method for a a trace session hides the EVENT_TRACE_LOGFILE that is modified by OpenTrace. This structure is occasionally needed so that the trace consumer can interpret the trace data correctly. In particular, it indicates if timestamps are in raw format or not.

Provided an alternative to start() that allows to open() session and to process() the session in separate calls. Open returns EVENT_TRACE_LOGFILE and does everything start does except for the call to ProcessTrace