Add unit tests to cover parsing string property at end of buffer
Modified the record_generator to trim the generated record (optionally) if the last property is a string.
Include a test for UNICODE strings, but do not have an ETW event for to test the ANSI string case.
The included test results in error, when the size check changes are reverted.