Improve property parsing (#41)
* Parse each property once

The last property parsed was not added to the cache and the property
index was not advanced. This resulted in extra work during parsing and
the same property being present multiple times in the cache.

* Compute the size of strings correctly

Strings at the end of a event buffer may not contain a null terminator.

* Replace const wchar_t* with auto

* Add unit tests to cover parsing string property at end of buffer

Modified the record_generator to trim the generated record (optionally) if the last property is a string.
Include a test for UNICODE strings, but do not have an ETW event for to test the ANSI string case.
The included test results in error, when the size check changes are reverted.