Provide separate open() and process() session calls (#54)
* Parse each property once

The last property parsed was not added to the cache and the property
index was not advanced. This resulted in extra work during parsing and
the same property being present multiple times in the cache.

* Compute the size of strings correctly

Strings at the end of a event buffer may not contain a null terminator.

* Provide separate open and process session calls

The current start() method for a a trace session hides the EVENT_TRACE_LOGFILE that is modified by OpenTrace. This structure is occasionally needed so that the trace consumer can interpret the trace data correctly. In particular, it indicates if timestamps are in raw format or not.

Provided an alternative to start() that allows to open() session and to process() the session in separate calls. Open returns EVENT_TRACE_LOGFILE and does everything start does except for the call to ProcessTrace

* Replace const wchar_t* with auto

* Add unit tests to cover parsing string property at end of buffer

Modified the record_generator to trim the generated record (optionally) if the last property is a string.
Include a test for UNICODE strings, but do not have an ETW event for to test the ANSI string case.
The included test results in error, when the size check changes are reverted.