Replace "min" by "std::min<size_t>" to support NOMINMAX
Including krabsetw from a project that defines "NOMINMAX" would cause a
compilation error due to the absence of the "min" macro. The issue is
fixed by using "std::min<size_t>" instead  of "min". Note that we
explicitly specify the template type in "std::min<size_t>" to avoid a
clash with the "min" macro in the case where "NOMINMAX" is not defined.

Also, check that there is no truncation when casting size_t to USHORT.

Tests: ran all test