Add optional trace::create() method
Some providers are configured through the use of TraceSetInformation
which takes in a session handle. The session handle is available after
the call to StartTrace in register_trace(). Add a "create()" method that
calls "register_trace" independently from "start()" to give an
opportunity to use "TraceSetInformation" whith the returned handle
before the blocking call to start(). Note that "create()" is optional.
If not called, the trace session will be created automatically when
calling "start()", as before.

Note that we could have used "ControlTrace" to obtain the handle of an
existing trace. However, since the session handle was assigned through
the start() method, which runs in its own thread, there is a race
condition.

Tests: all tests pass.